+++ pkg/gofr/datasource/pubsub/kafka/kafka.go
@@ -2,7 +2,6 @@ package kafka
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"sync"
 	"time"
@@ -176,46 +175,4 @@ func (k *kafkaClient) getNewReader(topic string) Reader {
 	})
 
 	return reader
-}
-
-func (k *kafkaClient) getReaderStatsAsMap() []interface{} {
-	readerStats := make([]interface{}, 0)
-
-	for _, reader := range k.reader {
-		var readerStat map[string]interface{}
-		if err := convertStructToMap(reader.Stats(), &readerStat); err != nil {
-			k.logger.Errorf("Kafka Reader Stats processing failed: %v", err)
-			continue // Log the error but continue processing other readers
-		}
-
-		readerStats = append(readerStats, readerStat)
-	}
-
-	return readerStats
-}
-
-func (k *kafkaClient) getWriterStatsAsMap() map[string]interface{} {
-	writerStats := make(map[string]interface{})
-
-	if err := convertStructToMap(k.writer.Stats(), &writerStats); err != nil {
-		k.logger.Errorf("Kafka Writer Stats processing failed: %v", err)
-		return nil
-	}
-
-	return writerStats
-}
-
-// convertStructToMap tries to convert any struct to a map representation by first marshaling it to JSON, then unmarshalling into a map.
-func convertStructToMap(input, output interface{}) error {
-	body, err := json.Marshal(input)
-	if err != nil {
-		return err
-	}
-
-	err = json.Unmarshal(body, &output)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
